Refacotr RateService to separate project and add caching
