+++ csharp/src/Blockchain.Swap/Blockchain.Swap.csproj
@@ -11,5 +11,6 @@
 		<ProjectReference Include="..\Infrastructure.Logging.OpenTelemetry\Infrastructure.Logging.OpenTelemetry.csproj" />
 		<ProjectReference Include="..\Infrastructure.Secret.HashicorpKeyVault\Infrastructure.Secret.HashicorpKeyVault.csproj" />
 	  	<ProjectReference Include="..\Infrastructure.TokenPrice.Coingecko\Infrastructure.TokenPrice.Coingecko.csproj" />
+		<ProjectReference Include="..\Infrastructure.MarketMaker\Infrastructure.MarketMaker.csproj" />
 	</ItemGroup>
 </Project>
