+++ csharp/src/Blockchain.Swap/Program.cs
@@ -4,6 +4,7 @@ using Train.Solver.Infrastructure.TokenPrice.Coingecko;
 using Train.Solver.Data.Npgsql.Extensions;
 using Train.Solver.Infrastructure.Logging.OpenTelemetry;
 using Train.Solver.Infrastructure.Secret.HashicorpKeyVault;
+using Train.Solver.Infrastructure.MarketMaker;
 
 IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureAppConfiguration(builder =>
@@ -20,7 +21,8 @@ IHost host = Host.CreateDefaultBuilder(args)
             .WithNpgsqlRepositories()
             .WithCoreWorkflows()
             .WithHashicorpKeyVault()
-            .WithCoingeckoPrices();
+            .WithCoingeckoPrices()
+            .WithMarketMaker();
     })
     .Build();
