+++ csharp/src/Infrastructure.MarketMaker/Infrastructure.MarketMaker.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+		<AssemblyName>Train.Solver.Infrastructure.MarketMaker</AssemblyName>
+		<RootNamespace>Train.Solver.Infrastructure.MarketMaker</RootNamespace>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Binance.Net" Version="10.19.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Infrastructure.Abstractions\Infrastructure.Abstractions.csproj" />
+	  <ProjectReference Include="..\Infrastructure.DependencyInjection\Infrastructure.DependencyInjection.csproj" />
+	</ItemGroup>
+
+</Project>
