+++ csharp/src/Infrastructure/Infrastructure.csproj
@@ -6,7 +6,6 @@
   </PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Binance.Net" Version="10.19.1" />
 		<PackageReference Include="RedLock.net" Version="2.3.2" />
 		<PackageReference Include="Temporalio.Extensions.Hosting" Version="1.5.0" />
 		<PackageReference Include="Microsoft.Extensions.Http" Version="9.0.3" />
